% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_params}
\alias{check_params}
\title{check conformity of parameters for JSON section}
\usage{
check_params(test_params, template_params, tz)
}
\arguments{
\item{test_params}{list of parameters from input}

\item{template_params}{list of parameters hardcoded in package}

\item{tz}{timezone from input}
}
\value{
No return value, called for side effects
}
\description{
check conformity of parameters for JSON section
}
\examples{
\dontrun{
check_params(params,template_params,"UTC")
}
}
