% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{get_azuremaps_token}
\alias{get_azuremaps_token}
\title{Get Azure Maps API Authentication Token}
\usage{
get_azuremaps_token()
}
\value{
A character string containing the Azure Maps API token.
}
\description{
Retrieves the Azure Maps API token from the environment.
}
\examples{
\dontrun{
get_azuremaps_token()
}
}
