% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_response}
\alias{sample_response}
\title{A sample response obtained from the \code{get_route_directions} function}
\format{
a \code{httr2} response object
}
\description{
A sample response obtained from the \code{get_route_directions} function
}
\references{
Sample of the call based on the API documentation in \url{https://learn.microsoft.com/en-us/rest/api/maps/route/post-route-directions?view=rest-maps-2025-01-01&tabs=HTTP#examples}
}
\keyword{directions}
\keyword{response}
