% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_response_leeds}
\alias{sample_response_leeds}
\title{A sample response obtained from the \code{get_route_directions} function}
\format{
a \code{httr2} response object
}
\description{
A sample response obtained from the \code{get_route_directions} function
}
\keyword{directions}
\keyword{response}
