% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{set_azuremaps_token}
\alias{set_azuremaps_token}
\title{Saves an Authentication Token for the Azure Maps API}
\usage{
set_azuremaps_token(token)
}
\arguments{
\item{token}{a \code{string} with the token}
}
\value{
TRUE if token correctly set
}
\description{
Saves an Authentication Token for the Azure Maps API
}
\examples{

mytoken <- "ivRgw7ZAGFedfwIdASezecdnETZDsdETB4Bqv3pbs5X8JDNnt1pQtpxDmpR6as2k"
set_azuremaps_token(mytoken)

}
