% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_points.R
\name{get_point}
\alias{get_point}
\alias{get_point.default}
\alias{get_point.numeric}
\alias{get_point.matrix}
\alias{get_point.sf}
\alias{get_point.sfc}
\title{Create sfc Object from Coordinates or sf Object}
\usage{
get_point(x)

\method{get_point}{default}(x)

\method{get_point}{numeric}(x)

\method{get_point}{matrix}(x)

\method{get_point}{sf}(x)

\method{get_point}{sfc}(x)
}
\arguments{
\item{x}{A numeric vector of length 2, a matrix with coordinates, or an sf object of POINT type.}
}
\value{
An sfc object with coordinates in EPSG:4326.
}
\description{
Converts a pair of coordinates, a matrix of coordinates, or an sf POINT object into an sfc object for use in GeoJSON bodies.
}
\examples{
get_point(c(-122.201399,47.608678))
}
