% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{set_azuremaps_token}
\alias{set_azuremaps_token}
\title{Set Azure Maps API Authentication Token}
\usage{
set_azuremaps_token(token)
}
\arguments{
\item{token}{A character string containing the Azure Maps API token.}
}
\value{
Logical TRUE if the token is correctly set.
}
\description{
Saves an authentication token for the Azure Maps API in the environment.
}
\examples{
\dontrun{
set_azuremaps_token("your_token_here")
}
}
